Create latency form for generation
